python-backend/requirements.txt
# FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6          # File uploads
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1                   # Migrations

# File Processing
openpyxl==3.1.2                   # Excel (xlsx)
python-docx==1.1.0                # Word (docx)
mammoth==1.6.0                    # Word to HTML

# AI/ML
anthropic==0.18.1                 # Anthropic SDK
openai==1.3.7                     # OpenAI SDK (for OpenRouter)
httpx==0.25.2                     # Async HTTP client
langchain==0.1.0                  # LLM framework
langsmith==0.0.69                 # LangSmith telemetry
langgraph==0.0.50                 # Workflow orchestration

# NLP & Embeddings
spacy==3.7.2
nltk==3.8.1
sentence-transformers==2.7.0

# Message Queue
aio-pika==9.2.1                   # RabbitMQ async client
celery==5.3.4                     # Distributed tasks (optional)

# OAuth & Auth
authlib==1.2.1                    # OAuth2 client (Jira)
httpx-oauth==0.13.0               # Alternative OAuth
itsdangerous==2.1.2               # Session signing
python-jose[cryptography]==3.3.0  # JWT

# Utilities
python-dotenv==1.0.0
pydantic-extra-types==2.3.0
bcrypt==4.1.1                     # Password hashing

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0       # Error tracking

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2                     # Test client
pytest-cov==4.1.0
pytest-mock==3.12.0               # Mocking support
